+++ src/WriterXml.cs
@@ -58,7 +58,36 @@ namespace Dec
 
         public WriterNode StartDec(Type type, string decName)
         {
-            return WriterNodeXml.StartDec(this, decs, type.ComposeDecFormatted(), decName);
+            string typeName;
+            Type overrideClass = null;
+            if (type.IsGenericType)
+            {
+                // okay, this just got more complicated
+                // we need to find a dec subclass that doesn't have any generics, then attach an attribute
+                // this is . . . not great design, honestly, I'm not sure I like this format right now
+                // but, fine, it's ugly but it will work
+                Type baseType = type.BaseType;
+                while (baseType.IsGenericType)
+                {
+                    baseType = baseType.BaseType;
+                }
+
+                typeName = baseType.ComposeDecFormatted();
+                overrideClass = type;
+            }
+            else
+            {
+                typeName = type.ComposeDecFormatted();
+            }
+
+            var nodeXml = WriterNodeXml.StartDec(this, decs, typeName, decName);
+
+            if (overrideClass != null)
+            {
+                nodeXml.TagClass(overrideClass);
+            }
+
+            return nodeXml;
         }
 
         public string Finish(bool pretty)